python test maybe

+++ electron/cpHandler.js
@@ -22,10 +22,14 @@ async function getPreview(filename) {
 
 async function generateAndRunPy(filename) {
     try {
-        fs.unlinkSync(`./ipynb-files/${filename}.py`)
+        
+        if(fs.existsSync(`./ipynb-files/${filename}.py`)){
+            fs.unlinkSync(`./ipynb-files/${filename}.py`)
+        }
         
         const {stdout_convert, stderr_convert} = await execute(`ipython nbconvert --to script ./ipynb-files/${filename}.ipynb`)
         const {stdout, stderr} = await execute(`python ./ipynb-files/${filename}.py`)
+        
         // if(stderr) {
         //     console.log("inside strerr", stderr)
         //     throw stderr
